# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Movie Club API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /reactions:
    get:
      operationId: getReactions
      tags: [reactions]
      summary: Get reactions for an entity
      parameters:
        - name: entityType
          in: query
          required: true
          schema:
            type: string
        - name: entityId
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Reactions summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReactionsResponse"
    post:
      operationId: createReaction
      tags: [reactions]
      summary: Add a reaction to an entity
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateReactionBody"
      responses:
        "201":
          description: Reaction created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reaction"
  /reactions/toggle:
    post:
      operationId: toggleReaction
      tags: [reactions]
      summary: Toggle a reaction (add if not exists, remove if exists)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ToggleReactionBody"
      responses:
        "200":
          description: Reaction toggled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ToggleReactionResponse"
  /reactions/details:
    get:
      operationId: getReactionDetails
      tags: [reactions]
      summary: Get detailed reactions (who reacted with what)
      parameters:
        - name: entityType
          in: query
          required: true
          schema:
            type: string
        - name: entityId
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Detailed reactions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReactionDetailsResponse"
  /reactions/{reactionId}:
    delete:
      operationId: deleteReaction
      tags: [reactions]
      summary: Remove a reaction
      parameters:
        - name: reactionId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Reaction removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
...
    Reaction:
      type: object
      properties:
        id:
          type: integer
        entityType:
          type: string
        entityId:
          type: integer
        stickerId:
          type: integer
        userId:
          type: integer
        createdAt:
          type: string
      required:
        - id
        - entityType
        - entityId
        - stickerId
        - userId
        - createdAt
    CreateReactionBody:
      type: object
      properties:
        entityType:
          type: string
        entityId:
          type: integer
        stickerId:
          type: integer
      required:
        - entityType
        - entityId
        - stickerId
    ToggleReactionBody:
      type: object
      properties:
        entityType:
          type: string
        entityId:
          type: integer
        stickerId:
          type: integer
      required:
        - entityType
        - entityId
        - stickerId
    ToggleReactionResponse:
      type: object
      properties:
        added:
          type: boolean
        reactionId:
          type: integer
          nullable: true
      required:
        - added
    ReactionSummary:
      type: object
      properties:
        stickerId:
          type: integer
        stickerName:
          type: string
        stickerImageUrl:
          type: string
        count:
          type: integer
        userReacted:
          type: boolean
      required:
        - stickerId
        - stickerName
        - stickerImageUrl
        - count
        - userReacted
    ReactionsResponse:
      type: object
      properties:
        reactions:
          type: array
          items:
            $ref: "#/components/schemas/ReactionSummary"
        totalCount:
          type: integer
      required:
        - reactions
        - totalCount
    ReactionDetail:
      type: object
      properties:
        id:
          type: integer
        stickerId:
          type: integer
        stickerName:
          type: string
        stickerImageUrl:
          type: string
        userId:
          type: integer
        username:
          type: string
        createdAt:
          type: string
      required:
        - id
        - stickerId
        - stickerName
        - stickerImageUrl
        - userId
        - username
        - createdAt
    ReactionDetailsResponse:
      type: object
      properties:
        reactions:
          type: array
          items:
            $ref: "#/components/schemas/ReactionDetail"
      required:
        - reactions