# lib/api-zod/src/generated/types/reactionDetail.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Movie Club API
 * OpenAPI spec version: 0.1.0
 */

export interface ReactionDetail {
  id: number;
  stickerId: number;
  stickerName: string;
  stickerImageUrl: string;
  userId: number;
  username: string;
  createdAt: string;
}