# lib/api-zod/src/generated/types/reactionDetailsResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Movie Club API
 * OpenAPI spec version: 0.1.0
 */
import type { ReactionDetail } from "./reactionDetail";

export interface ReactionDetailsResponse {
  reactions: ReactionDetail[];
}