# lib/api-zod/src/generated/types/reactionSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Movie Club API
 * OpenAPI spec version: 0.1.0
 */

export interface ReactionSummary {
  stickerId: number;
  stickerName: string;
  stickerImageUrl: string;
  count: number;
  userReacted: boolean;
}