# lib/api-zod/src/generated/types/reactionsResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Movie Club API
 * OpenAPI spec version: 0.1.0
 */
import type { ReactionSummary } from "./reactionSummary";

export interface ReactionsResponse {
  reactions: ReactionSummary[];
  totalCount: number;
}