# lib/api-zod/src/generated/types/toggleReactionBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Movie Club API
 * OpenAPI spec version: 0.1.0
 */

export interface ToggleReactionBody {
  entityType: string;
  entityId: number;
  stickerId: number;
}