# lib/api-zod/src/generated/types/toggleReactionResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Movie Club API
 * OpenAPI spec version: 0.1.0
 */

export interface ToggleReactionResponse {
  added: boolean;
  reactionId?: number | null;
}